requirements.txt
# MLA V3 依赖包

# 核心依赖
litellm>=1.0.0          # 统一的LLM接口
pyyaml>=6.0             # YAML配置文件解析
tiktoken

